Update script.js
My idea is too make a third button on the startup screen for theme selection and when the button is pressed it would go to a different page where the user can change what theme they would want. I made an attempt to make a button first I added it in and commented the lines of code.